Remove unused functions
