+++ libraries/math_expression/CMakeLists.txt
@@ -2,7 +2,7 @@ cmake_minimum_required(VERSION 3.16)
 
 project(math_expression)
 
-add_library(math_expression_lib STATIC utils.cpp tokenizer.cpp)
+add_library(math_expression_lib STATIC utils.cpp)
 target_include_directories(
     math_expression_lib
     PUBLIC #
