+++ libraries/math_expression/utils.cpp
@@ -1,57 +1,193 @@
 #include "utils.hpp"
 
-#include <algorithm>
-#include <numeric>
-#include <ranges>
+#include <bitset>
+#include <cctype>
+#include <iomanip>
 #include <stdexcept>
-#include <vector>
+#include <unordered_map>
 
 namespace {
-
-void variables_format_check(const std::vector<std::string>& variables) {
-    for(const std::string& var : variables) 
-        if (!var.empty() && !std::isalpha(var.front()))
-            throw std::domain_error{"Wrong variables format <" + var + ">. "
-                                    "Variables must start with latin letter, variable cannot start with a number."};
+constexpr auto make_lut(std::string_view chars) {
+    std::bitset<256> lut{};
+    for (const auto c : chars) lut[static_cast<uint8_t>(c)] = true;
+    return lut;
 }
-
-}
+constexpr auto whitespace_chars = make_lut(" \t\n\r\v\f");
+constexpr auto integer_chars = make_lut("0123456789");
+constexpr auto float_chars = make_lut("0123456789.");
+constexpr auto operator_chars = make_lut("+-*/^~");
+constexpr auto symbol_chars = make_lut("abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ_0123456789.");
+constexpr auto separator_chars = make_lut(",:|");
+constexpr auto parenthesis_chars = make_lut("()");
+}  // namespace
 
 namespace formula::utils {
 
-std::string trim(std::string str)
-{
-    str.erase(str.find_last_not_of(' ')+1);    //suffixing spaces
-    str.erase(0, str.find_first_not_of(' '));  //prefixing spaces
-    return str;
+std::ostream& operator<<(std::ostream& os, token_t token) {
+    // ParenthesisRight - 16 chars
+    os << '[' << std::setw(16);
+    switch (token.type) {
+        case token_t::type_t::Unknown:
+            os << "Unknown";
+            break;
+        case token_t::type_t::Number:
+            os << "Number";
+            break;
+        case token_t::type_t::ParenthesisLeft:
+            os << "ParenthesisLeft";
+            break;
+        case token_t::type_t::ParenthesisRight:
+            os << "ParenthesisRight";
+            break;
+        case token_t::type_t::Separator:
+            os << "Separator";
+            break;
+        case token_t::type_t::Operator:
+            os << "Operator";
+            break;
+        case token_t::type_t::Symbol:
+            os << "Symbol";
+            break;
+    }
+    os << "] : " << token.str;
+
+    return os;
 }
 
-std::string& delete_all(std::string& str, char symb) {
-    str.erase(std::remove(str.begin(), str.end(), symb), str.end());
-    return str;
+std::vector<token_t> tokenize(std::string_view input) {
+    constexpr std::string_view trim_chars = " \t\n\r\v\f";
+    const auto first = input.find_first_not_of(trim_chars);
+    if (first == std::string_view::npos) return {};
+    const auto last = input.find_last_not_of(trim_chars);
+    input = input.substr(first, last - first + 1);
+
+    std::vector<token_t> tokens;
+
+    enum class state_t : uint8_t {
+        NewToken,
+        IntegerNumber,
+        RealNumber,
+        Symbol,
+        ParenthesisLeft,
+        ParenthesisRight,
+        Separator,
+        Operator,
+        CompleteToken,
+        Terminate
+    };
+    state_t state = state_t::NewToken;
+
+    token_t::type_t token_type = token_t::type_t::Unknown;
+    size_t parenthesis_checker = 0;
+    std::string_view::const_iterator token_begin = input.end();
+
+    for (auto it = input.begin(); state != state_t::Terminate;) {
+        if(static_cast<uint8_t>(*it) > 127)
+        throw std::domain_error{"Wrong expression format. The expression contains non-ANSI characters."};
+        // If the iterator is at the end of the input we use '\0' as a sentinel value.
+        const char c = (it == input.end()) ? '\0' : *it;
+        // Handle incorrect implicit conversion. I think so... In any case, the app fails without this cast.
+        uint8_t idx = static_cast<uint8_t>(c);
+
+        switch (state) {
+            case state_t::NewToken: {
+                token_type = token_t::type_t::Unknown;
+                token_begin = it;
+                if (whitespace_chars[idx]) {
+                    state = state_t::NewToken;
+                } else if (float_chars[idx]) {
+                    state = state_t::IntegerNumber;
+                    continue;
+                } else if (operator_chars[idx]) {
+                    state = state_t::Operator;
+                    continue;
+                } else if (parenthesis_chars[idx]) {
+                    state = c == '(' ? state_t::ParenthesisLeft : state_t::ParenthesisRight;
+                    continue;
+                } else if (separator_chars[idx]) {
+                    state = state_t::Separator;
+                    continue;
+                } else {  // if nothing detected can be a symbol
+                    state = state_t::Symbol;
+                    continue;
+                }
+            } break;
+            case state_t::IntegerNumber: {
+                if (!float_chars[idx] && symbol_chars[idx]) {
+                    throw std::domain_error{"Wrong expression format. The expression contains invalid numeric construction"};
+                } else if (!float_chars[idx]) {
+                    token_type = token_t::type_t::Number;
+                    state = state_t::CompleteToken;
+                    continue;
+                } else {
+                    if (c == '.') state = state_t::RealNumber;
+                }
+            } break;
+            case state_t::RealNumber: {
+                if (!integer_chars[idx] && symbol_chars[idx]) {
+                    throw std::domain_error{"Wrong expression format. The expression contains invalid dots. Dots are only allowed in number representation"};
+                } else if (!integer_chars[idx]) {
+                    token_type = token_t::type_t::Number;
+                    state = state_t::CompleteToken;
+                    continue;
+                }
+            } break;
+            case state_t::Operator: {
+                token_type = token_t::type_t::Operator;
+                state = state_t::CompleteToken;
+            } break;
+            case state_t::ParenthesisLeft: {
+                ++parenthesis_checker;
+                token_type = token_t::type_t::ParenthesisLeft;
+                state = state_t::CompleteToken;
+            } break;
+            case state_t::ParenthesisRight: {
+                --parenthesis_checker;
+                token_type = token_t::type_t::ParenthesisRight;
+                state = state_t::CompleteToken;
+            } break;
+            case state_t::Separator: {
+                token_type = token_t::type_t::Separator;
+                state = state_t::CompleteToken;
+            } break;
+            case state_t::Symbol: {
+                if (!symbol_chars[idx]) {
+                    token_type = token_t::type_t::Symbol;
+                    state = state_t::CompleteToken;
+                    continue;
+                }
+            } break;
+            case state_t::CompleteToken: {
+                if(token_begin == it)
+                    throw std::runtime_error{"Wrong tokenizer behaviour. Empty token parsed."};
+                tokens.emplace_back(token_t{token_type, {token_begin, it}});
+                state = it == input.end() ? state_t::Terminate : state_t::NewToken;
+                continue;
+            } break;
+        }
+        if(it != input.end())
+            ++it;
+    }
+
+    if (parenthesis_checker)
+        throw std::domain_error{"Wrong expression format. The expression contains open parentheses."};
+    return tokens;
 }
 
-std::size_t count_all(const std::string& s, char symb) {
-    return std::accumulate(s.begin(), s.end(), std::size_t(0), [&](std::size_t res, char a){ return res + std::size_t(a == symb); });
+std::unordered_map<std::string, std::size_t> get_variables(const std::span<token_t>& tokens) {
+    std::unordered_map<std::string, std::size_t> variables;
+
+    for (const auto& token : tokens) {
+        if (token.type == token_t::type_t::Symbol) {
+            variables[token.str] = variables.size();
+            continue;
+        }
+        // numbers/operators/parentheses are invalid in variables declaration
+        throw std::domain_error{"Wrong variables format. Variables must be declared as symbols before ':' separator."};
+    }
+
+    return variables;
 }
-
-bool is_latin_str(const std::string& s) {
-    if (s.size() == 0 || s.size() > 1)
-        throw std::domain_error{"Wrong format, latyn symbol contains of one symbol."};
-    for(const char d : s)
-        if (!std::isalpha(d)) 
-            return false;
-    return true;
-};
-
-bool is_number(const std::string& s) {
-    if (const std::size_t dots = count_all(s, '.'); dots > 1)
-        return false;
-    for(const char d : s)
-        if (d != '.' && !std::isdigit(d)) 
-            return false;
-    return true;
-};
 
 const std::unordered_map<std::string, std::size_t>& get_operator_priority() {
     static const std::unordered_map<std::string, size_t> operator_priority{
@@ -69,10 +205,6 @@ const std::unordered_map<std::string, std::size_t>& get_operator_priority() {
     return operator_priority;
 }
 
-const std::unordered_set<char>& get_one_sym_operators() {
-    static const std::unordered_set<char> one_sym_operators{'(', ')', '+', '-', '*', '/', '~', '^'};
-    return one_sym_operators;
-}
 
 void check_variables_admissibility(const std::unordered_map<std::string, std::size_t>& variables) {
     const auto& operator_priority = get_operator_priority();
@@ -80,38 +212,6 @@ void check_variables_admissibility(const std::unordered_map<std::string, std::si
         if (operator_priority.contains(variable))
             throw std::domain_error{"Invalid variable designation. Variable name <" + variable + "> \
                                         is unavailable."};   
-}
-
-void parentheses_check(const std::string& infix_notation) {
-    if (formula::utils::count_all(infix_notation, '(') != formula::utils::count_all(infix_notation, ')'))
-        throw std::domain_error{"Wrong expression format. The expression contains open parentheses."};
-}
-
-void dots_check(const std::string& infix_notation) {
-    static constexpr char dot = '.';
-    if (infix_notation[0] == dot || infix_notation[infix_notation.size() - 1] == dot)
-        throw std::domain_error{"Wrong expression format. The expression can not be started or finished with dot. Dots are only allowed in number representation."};
-    for(std::size_t position = infix_notation.find(dot); position != std::string::npos; position = infix_notation.find(dot, position + 1)) {
-        if (!std::isdigit(infix_notation[position + 1]))
-            throw std::domain_error{"Wrong expression format. The expression contains invalid dots. Dots are only allowed in number representation"};
-    }
-}
-
-std::unordered_map<std::string, std::size_t> get_variables(std::string pre_variables) {
-    static constexpr char delimiter_symbol = ' ';
-    std::vector<std::string> variables = {};
-    std::size_t delimiter = pre_variables.find(delimiter_symbol);   
-    while (delimiter != std::string::npos || pre_variables.size() > 0) {
-        variables.push_back(pre_variables.substr(0, delimiter));
-        pre_variables = delimiter < pre_variables.size() ? pre_variables.substr(delimiter + 1) : "";
-        delimiter = pre_variables.find(delimiter_symbol);
-    }
-    variables_format_check(variables);
-    std::unordered_map<std::string, std::size_t> res;
-    std::ranges::for_each(std::views::iota(size_t(0), variables.size()), [&res, &variables](std::size_t i) { 
-            res[variables[i]] = i; 
-    });
-    return res;
 }
 
 }
